tests for the new flow
