+++ test/donations.routes.test.js
@@ -7,13 +7,13 @@ const expect = require('chai').expect;
 const request = require('supertest');
 const sinon = require('sinon');
 const nock = require('nock');
-const chance = require('chance').Chance();
 
 const utils = require('../test/utils.js')();
 const generatePlanId = require('../server/lib/utils.js').planId;
 const roles = require('../server/constants/roles');
 const constants = require('../server/constants/transactions');
 const emailLib = require('../server/lib/email');
+const donationLib = require('../server/lib/donation');
 
 /**
  * Variables.
@@ -40,17 +40,11 @@ describe('donations.routes.test.js', () => {
   var group;
   var group2;
   var nocks = {};
-  var stripeEmail;
   var sandbox = sinon.sandbox.create();
 
-  var stubStripe = () => {
-    var mock = stripeMock.accounts.create;
-    mock.email = chance.email();
-    stripeEmail = mock.email;
-
-    var stub = sinon.stub(app.stripe.accounts, 'create');
-    stub.yields(null, mock);
-  };
+  beforeEach(() => {
+    sandbox.stub(donationLib, 'processDonation');
+  })
 
   beforeEach(() => utils.cleanAllDb().tap(a => application = a));
 
@@ -62,24 +56,6 @@ describe('donations.routes.test.js', () => {
 
   // Create a user.
   beforeEach(() => models.User.create(userData).tap(u => user = u));
-
-  // Nock for customers.create.
-  beforeEach(() => {
-    nocks['customers.create'] = nock(STRIPE_URL)
-      .post('/v1/customers')
-      .reply(200, stripeMock.customers.create);
-  });
-
-  // Nock for retrieving balance transaction
-  beforeEach(() => {
-    nocks['balance.retrieveTransaction'] = nock(STRIPE_URL)
-      .get('/v1/balance/history/txn_165j8oIqnMN1wWwOKlPn1D4y')
-      .reply(200, stripeMock.balance);
-  });
-
-  beforeEach(() => {
-    stubStripe();
-  });
 
   // Create a group.
   beforeEach((done) => {
@@ -103,11 +79,6 @@ describe('donations.routes.test.js', () => {
       });
   });
 
-  beforeEach(() => {
-    app.stripe.accounts.create.restore();
-    stubStripe();
-  });
-
   // Create a second group.
   beforeEach((done) => {
     request(app)
@@ -128,10 +99,6 @@ describe('donations.routes.test.js', () => {
           })
           .catch(done);
       });
-  });
-
-  beforeEach(() => {
-    app.stripe.accounts.create.restore();
   });
 
   beforeEach((done) => {
@@ -160,29 +127,6 @@ describe('donations.routes.test.js', () => {
     .tap(a => application2 = a)
     .then(() => application2.addGroup(group2)));
 
-  // Nock for charges.create.
-  beforeEach(() => {
-    var params = [
-      'amount=' + CHARGE * 100,
-      'currency=' + CURRENCY,
-      'customer=' + stripeMock.customers.create.id,
-      'description=' + encodeURIComponent(`OpenCollective: ${group.slug}`),
-      'application_fee=54',
-      encodeURIComponent('metadata[groupId]') + '=' + group.id,
-      encodeURIComponent('metadata[groupName]') + '=' + encodeURIComponent(groupData.name),
-      encodeURIComponent('metadata[customerEmail]') + '=' + encodeURIComponent(user.email),
-      encodeURIComponent('metadata[paymentMethodId]') + '=1'
-    ].join('&');
-
-    nocks['charges.create'] = nock(STRIPE_URL)
-      .post('/v1/charges', params)
-      .reply(200, stripeMock.charges.create);
-  });
-
-  afterEach(() => {
-    nock.cleanAll();
-  });
-
   afterEach(() => {
     utils.clearbitStubAfterEach(sandbox);
   });
@@ -210,10 +154,6 @@ describe('donations.routes.test.js', () => {
           .end(done);
       });
 
-      it('successfully creates a Stripe customer', () => {
-        expect(nocks['customers.create'].isDone()).to.be.true;
-      });
-
       it('successfully creates a paymentMethod with the UserId', (done) => {
         models.PaymentMethod
           .findAndCountAll({})
@@ -222,18 +162,9 @@ describe('donations.routes.test.js', () => {
             expect(res.rows[0]).to.have.property('UserId', user.id);
             expect(res.rows[0]).to.have.property('token', STRIPE_TOKEN);
             expect(res.rows[0]).to.have.property('service', 'stripe');
-            expect(res.rows[0]).to.have.property('customerId', stripeMock.customers.create.id);
             done();
           })
           .catch(done);
-      });
-
-      it('successfully makes a Stripe charge', () => {
-        expect(nocks['charges.create'].isDone()).to.be.true;
-      });
-
-      it('successfully gets a Stripe balance', () => {
-        expect(nocks['balance.retrieveTransaction'].isDone()).to.be.true;
       });
 
       it('successfully creates a donation in the database', (done) => {
@@ -247,31 +178,6 @@ describe('donations.routes.test.js', () => {
             expect(res.rows[0]).to.have.property('amount', CHARGE*100);
             expect(res.rows[0]).to.have.property('title',
               `Donation to ${group.name}`);
-            done();
-          })
-          .catch(done);
-      });
-
-      it('successfully creates a transaction in the database', (done) => {
-        models.Transaction
-          .findAndCountAll({})
-          .then((res) => {
-            expect(res.count).to.equal(1);
-            expect(res.rows[0]).to.have.property('UserId', user.id);
-            expect(res.rows[0]).to.have.property('PaymentMethodId', 1);
-            expect(res.rows[0]).to.have.property('currency', CURRENCY);
-            expect(res.rows[0]).to.have.property('type', constants.type.DONATION);
-            expect(res.rows[0]).to.have.property('amount', CHARGE);
-            expect(res.rows[0]).to.have.property('amountInTxnCurrency', 1400); // taken from stripe mocks
-            expect(res.rows[0]).to.have.property('txnCurrency', 'USD');
-            expect(res.rows[0]).to.have.property('hostFeeInTxnCurrency', 0);
-            expect(res.rows[0]).to.have.property('platformFeeInTxnCurrency', 70);
-            expect(res.rows[0]).to.have.property('paymentProcessorFeeInTxnCurrency', 155);
-            expect(res.rows[0]).to.have.property('txnCurrencyFxRate', 0.785);
-            expect(res.rows[0]).to.have.property('netAmountInGroupCurrency', 922)
-            expect(res.rows[0].tags[0]).to.be.equal('Donation');
-            expect(res.rows[0]).to.have.property('description',
-              'Donation to ' + group.name);
             done();
           })
           .catch(done);
@@ -299,39 +205,6 @@ describe('donations.routes.test.js', () => {
           .end(done);
       });
 
-      // New nock for customers.create.
-      beforeEach(() => {
-        nocks['customers.create2'] = nock(STRIPE_URL)
-          .post('/v1/customers')
-          .reply(200, stripeMock.customers.create);
-      });
-
-      // Nock for charges.create.
-      beforeEach(() => {
-        var params = [
-          'amount=' + CHARGE2 * 100,
-          'currency=' + CURRENCY,
-          'customer=' + stripeMock.customers.create.id,
-          'description=' + encodeURIComponent(`OpenCollective: ${group.slug}`),
-          'application_fee=9',
-          encodeURIComponent('metadata[groupId]') + '=' + group.id,
-          encodeURIComponent('metadata[groupName]') + '=' + encodeURIComponent(group.name),
-          encodeURIComponent('metadata[customerEmail]') + '=' + encodeURIComponent(user.email),
-          encodeURIComponent('metadata[paymentMethodId]') + '=1'
-        ].join('&');
-
-        nocks['charges.create2'] = nock(STRIPE_URL)
-          .post('/v1/charges', params)
-          .reply(200, stripeMock.charges.create);
-      });
-
-      // Nock for retrieving balance transaction
-      beforeEach(() => {
-        nocks['balance.retrieveTransaction'] = nock(STRIPE_URL)
-          .get('/v1/balance/history/txn_165j8oIqnMN1wWwOKlPn1D4y')
-          .reply(200, stripeMock.balance);
-      });
-
       beforeEach((done) => {
         request(app)
           .post('/groups/' + group.id + '/payments')
@@ -348,10 +221,6 @@ describe('donations.routes.test.js', () => {
           .end(done);
       });
 
-      it('does not re-create a Stripe Customer with a same token', () => {
-        expect(nocks['customers.create2'].isDone()).to.be.false;
-      });
-
       it('does not re-create a paymentMethod', (done) => {
         models.PaymentMethod
           .findAndCountAll({})
@@ -362,138 +231,12 @@ describe('donations.routes.test.js', () => {
           .catch(done);
       });
 
-      it('successfully makes a new Stripe charge', () => {
-        expect(nocks['charges.create2'].isDone()).to.be.true;
-      });
-
       it('successfully creates a donation in the database', (done) => {
         models.Donation
           .findAndCountAll({})
           .then((res) => {
             expect(res.count).to.equal(2);
             expect(res.rows[1]).to.have.property('amount', CHARGE2*100);
-            done();
-          })
-          .catch(done);
-      });
-
-      it('successfully gets a Stripe balance', () => {
-        expect(nocks['balance.retrieveTransaction'].isDone()).to.be.true;
-      });
-
-      it('successfully creates a new transaction', (done) => {
-        models.Transaction
-          .findAndCountAll({order: 'id'})
-          .then((res) => {
-            expect(res.count).to.equal(2);
-            expect(res.rows[1]).to.have.property('amount', CHARGE2);
-            done();
-          })
-          .catch(done);
-      });
-
-    });
-
-    describe('Payment success by a user that is not part of the group yet', () => {
-
-      // Nock for charges.create.
-      beforeEach(() => {
-        var params = [
-          'amount=' + CHARGE * 100,
-          'currency=' + CURRENCY,
-          'customer=' + stripeMock.customers.create.id,
-          'description=' + encodeURIComponent(`OpenCollective: ${group2.slug}`),
-          'application_fee=54',
-          encodeURIComponent('metadata[groupId]') + '=' + group2.id,
-          encodeURIComponent('metadata[groupName]') + '=' + encodeURIComponent(group2.name),
-          encodeURIComponent('metadata[customerEmail]') + '=' + encodeURIComponent(EMAIL),
-          encodeURIComponent('metadata[paymentMethodId]') + '=1'
-        ].join('&');
-
-        nocks['charges.create'] = nock(STRIPE_URL)
-          .post('/v1/charges', params)
-          .reply(200, stripeMock.charges.create);
-      });
-
-      beforeEach((done) => {
-        request(app)
-          .post('/groups/' + group2.id + '/payments')
-          .send({
-            api_key: application.api_key,
-            payment: {
-              stripeToken: STRIPE_TOKEN,
-              amount: CHARGE,
-              currency: CURRENCY,
-              email: EMAIL
-            }
-          })
-          .expect(200)
-          .end(done);
-      });
-
-      it('successfully adds the user to the group as a backer', (done) => {
-        group2
-          .getUsers()
-          .then((users) => {
-            expect(users).to.have.length(2);
-            var backer = _.find(users, {email: EMAIL});
-            expect(backer.UserGroup.role).to.equal(roles.BACKER);
-            done();
-          })
-          .catch(done);
-      });
-
-    });
-
-    describe('Payment success by a user who is a MEMBER of the group and should become BACKER', () => {
-
-      // Add a user as a MEMBER
-      beforeEach(() => models.User.create(utils.data('user4'))
-        .tap(u => user4 = u)
-        .then(() => group2.addUserWithRole(user4, roles.MEMBER)));
-
-      // Nock for charges.create.
-      beforeEach(() => {
-        var params = [
-          'amount=' + CHARGE * 100,
-          'currency=' + CURRENCY,
-          'customer=' + stripeMock.customers.create.id,
-          'description=' + encodeURIComponent(`OpenCollective: ${group2.slug}`),
-          'application_fee=54',
-          encodeURIComponent('metadata[groupId]') + '=' + group2.id,
-          encodeURIComponent('metadata[groupName]') + '=' + encodeURIComponent(group2.name),
-          encodeURIComponent('metadata[customerEmail]') + '=' + encodeURIComponent(user4.email),
-          encodeURIComponent('metadata[paymentMethodId]') + '=1'
-        ].join('&');
-
-        nocks['charges.create'] = nock(STRIPE_URL)
-          .post('/v1/charges', params)
-          .reply(200, stripeMock.charges.create);
-      });
-
-      beforeEach((done) => {
-        request(app)
-          .post('/groups/' + group2.id + '/payments')
-          .set('Authorization', 'Bearer ' + user4.jwt(application2))
-          .send({
-            payment: {
-              stripeToken: STRIPE_TOKEN,
-              amount: CHARGE,
-              currency: CURRENCY,
-              email: user4.email
-            }
-          })
-          .expect(200)
-          .end(done);
-      });
-
-      it('successfully adds the user to the group as a backer', (done) => {
-        group2
-          .getUsers()
-          .then((users) => {
-            expect(users).to.have.length(3);
-            var backer = _.find(users, {email: user4.email});
-            expect(backer.UserGroup.role).to.equal(roles.BACKER);
             done();
           })
           .catch(done);
@@ -516,27 +259,6 @@ describe('donations.routes.test.js', () => {
         email: userData.email
       };
 
-      // Nock for charges.create.
-      beforeEach(() => {
-        var params = [
-          'amount=' + CHARGE * 100,
-          'currency=' + CURRENCY,
-          'customer=' + stripeMock.customers.create.id,
-          'description=' + encodeURIComponent(`OpenCollective: ${group2.slug}`),
-          'application_fee=54',
-          encodeURIComponent('metadata[groupId]') + '=' + group2.id,
-          encodeURIComponent('metadata[groupName]') + '=' + encodeURIComponent(group2.name),
-          encodeURIComponent('metadata[customerEmail]') + '=' + encodeURIComponent(userData.email),
-          encodeURIComponent('metadata[paymentMethodId]') + '=1'
-        ].join('&');
-
-        nocks['charges.create'] = nock(STRIPE_URL)
-          .post('/v1/charges', params)
-          .reply(200, stripeMock.charges.create);
-      });
-
-      beforeEach(() => sinon.spy(emailLib, 'send'));
-
       beforeEach('successfully makes a anonymous payment', (done) => {
         request(app)
           .post('/groups/' + group2.id + '/payments')
@@ -551,13 +273,6 @@ describe('donations.routes.test.js', () => {
           });
       });
 
-
-      afterEach(() => emailLib.send.restore());
-
-      it('successfully creates a Stripe customer', () => {
-        expect(nocks['customers.create'].isDone()).to.be.true;
-      });
-
       it('successfully creates a paymentMethod', (done) => {
         models.PaymentMethod
           .findAndCountAll({})
@@ -567,10 +282,6 @@ describe('donations.routes.test.js', () => {
             done();
           })
           .catch(done);
-      });
-
-      it('successfully makes a Stripe charge', () => {
-        expect(nocks['charges.create'].isDone()).to.be.true;
       });
 
       it('successfully creates a user', (done) => {
@@ -604,30 +315,6 @@ describe('donations.routes.test.js', () => {
           .catch(done);
       });
 
-      it('successfully gets a Stripe balance', () => {
-        expect(nocks['balance.retrieveTransaction'].isDone()).to.be.true;
-      });
-
-      it('successfully creates a transaction in the database', (done) => {
-        models.Transaction
-          .findAndCountAll({})
-          .then((res) => {
-            expect(res.count).to.equal(1);
-            expect(res.rows[0]).to.have.property('GroupId', group2.id);
-            expect(res.rows[0]).to.have.property('UserId', user.id);
-            expect(res.rows[0]).to.have.property('PaymentMethodId', 1);
-            expect(res.rows[0]).to.have.property('currency', CURRENCY);
-            expect(res.rows[0]).to.have.property('tags');
-            expect(res.rows[0]).to.have.property('payoutMethod', null);
-            expect(res.rows[0]).to.have.property('amount', data.amount);
-            done();
-          })
-          .catch(done);
-      });
-
-      it('successfully sends a thank you email', () => {
-        expect(emailLib.send.lastCall.args[1]).to.equal(userData.email.toLowerCase());
-      });
     });
 
     describe('Recurring payment success', () => {
@@ -646,160 +333,82 @@ describe('donations.routes.test.js', () => {
         email: EMAIL
       };
 
-      var customerId = stripeMock.customers.create.id;
-      var planId = generatePlanId({
-        currency: CURRENCY,
-        interval: data.interval,
-        amount: data.amount * 100
+
+      beforeEach((done) => {
+
+        request(app)
+          .post('/groups/' + group2.id + '/payments')
+          .send({
+            api_key: application2.api_key,
+            payment: data
+          })
+          .expect(200)
+          .end((e, res) => {
+            expect(e).to.not.exist;
+            done();
+          });
       });
 
-      var plan = _.extend({}, stripeMock.plans.create, {
-        amount: data.amount,
-        interval: data.interval,
-        name: planId,
-        id: planId
+      it('successfully creates a donation in the database', (done) => {
+        models.Donation
+          .findAndCountAll({})
+          .then((res) => {
+            expect(res.count).to.equal(1);
+            expect(res.rows[0]).to.have.property('UserId', 2);
+            expect(res.rows[0]).to.have.property('GroupId', group2.id);
+            expect(res.rows[0]).to.have.property('currency', CURRENCY);
+            expect(res.rows[0]).to.have.property('amount', data.amount*100);
+            expect(res.rows[0]).to.have.property('SubscriptionId');
+            expect(res.rows[0]).to.have.property('title',
+              `Donation to ${group2.name}`);
+            done();
+          })
+          .catch(done);
       });
 
-      beforeEach(() => {
-        nocks['plans.create'] = nock(STRIPE_URL)
-          .post('/v1/plans')
-          .reply(200, plan);
-        var params = [
-          `plan=${planId}`,
-          'application_fee_percent=5',
-          encodeURIComponent('metadata[groupId]') + '=' + group2.id,
-          encodeURIComponent('metadata[groupName]') + '=' + encodeURIComponent(group2.name),
-          encodeURIComponent('metadata[paymentMethodId]') + '=1',
-          encodeURIComponent('metadata[description]') + '=' + encodeURIComponent(`OpenCollective: ${group2.slug}`)
-        ].join('&');
-
-      nocks['subscriptions.create'] = nock(STRIPE_URL)
-        .post(`/v1/customers/${customerId}/subscriptions`, params)
-        .reply(200, stripeMock.subscriptions.create);
+      it('does not create a transaction', (done) => {
+        models.Transaction
+          .findAndCountAll({})
+          .then((res) => {
+            expect(res.count).to.equal(0);
+            done();
+          })
+          .catch(done);
       });
 
-      describe('plan does not exist', () => {
-        beforeEach((done) => {
+      it('creates a Subscription model', (done) => {
+        models.Subscription
+          .findAndCountAll({})
+          .then((res) => {
+            const subscription = res.rows[0];
 
-          nocks['plans.retrieve'] = nock(STRIPE_URL)
-            .get('/v1/plans/' + planId)
-            .reply(200, {
-              error: stripeMock.plans.create_not_found
-            });
-
-          request(app)
-            .post('/groups/' + group2.id + '/payments')
-            .send({
-              api_key: application2.api_key,
-              payment: data
-            })
-            .expect(200)
-            .end((e, res) => {
-              expect(e).to.not.exist;
-              done();
-            });
-        });
-
-        it('creates a plan if it doesn\'t exist', () => {
-          expect(nocks['plans.retrieve'].isDone()).to.be.true;
-          expect(nocks['plans.create'].isDone()).to.be.true;
-        });
-
+            expect(res.count).to.equal(1);
+            expect(subscription).to.have.property('amount', data.amount);
+            expect(subscription).to.have.property('interval', 'month');
+            expect(subscription).to.have.property('data');
+            expect(subscription).to.have.property('isActive', false);
+            expect(subscription).to.have.property('currency', CURRENCY);
+            done();
+          })
+          .catch(done);
       });
 
-      describe('plan exists', () => {
+      it('fails if the interval is not month or year', (done) => {
 
-        beforeEach((done) => {
-
-          nocks['plans.retrieve'] = nock(STRIPE_URL)
-            .get('/v1/plans/' + planId)
-            .reply(200, plan);
-
-          request(app)
-            .post('/groups/' + group2.id + '/payments')
-            .send({
-              api_key: application2.api_key,
-              payment: data
-            })
-            .expect(200)
-            .end((e, res) => {
-              expect(e).to.not.exist;
-              done();
-            });
-        });
-
-        it('uses the existing plan', () => {
-          expect(nocks['plans.create'].isDone()).to.be.false;
-          expect(nocks['plans.retrieve'].isDone()).to.be.true;
-        });
-
-        it('creates a subscription', () => {
-          expect(nocks['subscriptions.create'].isDone()).to.be.true;
-        });
-
-        it('successfully creates a donation in the database', (done) => {
-          models.Donation
-            .findAndCountAll({})
-            .then((res) => {
-              expect(res.count).to.equal(1);
-              expect(res.rows[0]).to.have.property('UserId', 2);
-              expect(res.rows[0]).to.have.property('GroupId', group2.id);
-              expect(res.rows[0]).to.have.property('currency', CURRENCY);
-              expect(res.rows[0]).to.have.property('amount', data.amount*100);
-              expect(res.rows[0]).to.have.property('SubscriptionId');
-              expect(res.rows[0]).to.have.property('title',
-                `Donation to ${group2.name}`);
-              done();
-            })
-            .catch(done);
-        });
-
-        it('does not create a transaction', (done) => {
-          models.Transaction
-            .findAndCountAll({})
-            .then((res) => {
-              expect(res.count).to.equal(0);
-              done();
-            })
-            .catch(done);
-        });
-
-
-        it('creates a Subscription model', (done) => {
-          models.Subscription
-            .findAndCountAll({})
-            .then((res) => {
-              const subscription = res.rows[0];
-
-              expect(res.count).to.equal(1);
-              expect(subscription).to.have.property('amount', data.amount);
-              expect(subscription).to.have.property('interval', plan.interval);
-              expect(subscription).to.have.property('stripeSubscriptionId', stripeMock.subscriptions.create.id);
-              expect(subscription).to.have.property('data');
-              expect(subscription).to.have.property('isActive', false);
-              expect(subscription).to.have.property('currency', CURRENCY);
-              done();
-            })
-            .catch(done);
-        });
-
-        it('fails if the interval is not month or year', (done) => {
-
-          request(app)
-            .post('/groups/' + group2.id + '/payments')
-            .send({
-              api_key: application2.api_key,
-              payment: _.extend({}, data, {interval: 'something'})
-            })
-            .expect(400, {
-              error: {
-                code: 400,
-                type: 'bad_request',
-                message: 'Interval should be month or year.'
-              }
-            })
-            .end(done);
-        });
+        request(app)
+          .post('/groups/' + group2.id + '/payments')
+          .send({
+            api_key: application2.api_key,
+            payment: _.extend({}, data, {interval: 'something'})
+          })
+          .expect(400, {
+            error: {
+              code: 400,
+              type: 'bad_request',
+              message: 'Interval should be month or year.'
+            }
+          })
+          .end(done);
       });
     });
 
@@ -1103,13 +712,6 @@ describe('donations.routes.test.js', () => {
 
     describe('Payment errors', () => {
 
-      beforeEach(() => {
-        nock.cleanAll();
-        nocks['customers.create'] = nock(STRIPE_URL)
-          .post('/v1/customers')
-          .replyWithError(stripeMock.customers.createError);
-      });
-
       it('fails if the accessToken contains live', (done) => {
         const payment = {
           stripeToken: STRIPE_TOKEN,
@@ -1134,27 +736,6 @@ describe('donations.routes.test.js', () => {
             .end(done);
         })
 
-      });
-
-      it('fails paying because of a paymentMethod declined', (done) => {
-        request(app)
-          .post('/groups/' + group.id + '/payments')
-          .set('Authorization', 'Bearer ' + user.jwt(application))
-          .send({
-            payment: {
-              stripeToken: STRIPE_TOKEN,
-              amount: CHARGE,
-              currency: CURRENCY
-            }
-          })
-          .expect(400)
-          .then(res => {
-            const error = res.body.error;
-            expect(error.message).to.equal('Your paymentMethod was declined');
-            expect(error.type).to.equal('StripePaymentMethodError');
-            expect(error.code).to.equal(400);
-            done();
-          });
       });
 
     });
