+++ test/group.model.test.js
@@ -8,7 +8,7 @@ const Transaction = models.Transaction;
 const Group = models.Group;
 const User = models.User;
 
-describe("Group model", () => {
+describe('Group model', () => {
 
   let group = {};
 
@@ -27,7 +27,7 @@ describe("Group model", () => {
   }];
 
   const transactions = [{
-    createdAt: new Date("2016-06-14"),
+    createdAt: new Date('2016-06-14'),
     amount: -10,
     netAmountInGroupCurrency: -1000,
     currency: 'USD',
@@ -35,7 +35,7 @@ describe("Group model", () => {
     description: 'pizza',
     tags: ['food']
   },{
-    createdAt: new Date("2016-07-14"),
+    createdAt: new Date('2016-07-14'),
     amount: -150,
     netAmountInGroupCurrency: -15000,
     currency: 'USD',
@@ -43,14 +43,14 @@ describe("Group model", () => {
     description: 'stickers',
     tags: ['marketing']
   },{
-    createdAt: new Date("2016-06-15"),
+    createdAt: new Date('2016-06-15'),
     amount: 250,
     netAmountInGroupCurrency: 25000,
     currency: 'USD',
     type: 'donation',
     UserId: 1
   },{
-    createdAt: new Date("2016-07-16"),
+    createdAt: new Date('2016-07-16'),
     amount: 500,
     netAmountInGroupCurrency: 50000,
     currency: 'USD',
@@ -58,7 +58,7 @@ describe("Group model", () => {
     UserId: 2
   }];
 
-  
+
   before((done) => utils.cleanAllDb().tap(a => done()));
 
   before((done) => {
@@ -68,12 +68,12 @@ describe("Group model", () => {
       .tap((g) => Transaction.createMany(transactions, { GroupId: group.id }))
       .then(() => done())
       .catch(e => {
-        console.error("Error in creating group", groupData, e, e.stack);
+        console.error('Error in creating group', groupData, e, e.stack);
         done();
       });
   });
 
-  it("computes the balance ", (done) => {
+  it('computes the balance ', (done) => {
     group.getBalance().then(balance => {
       let sum = 0;
       transactions.map(t => sum += t.netAmountInGroupCurrency);
@@ -81,9 +81,9 @@ describe("Group model", () => {
       done();
     })
   });
-  
-  it("computes the balance until a certain month", (done) => {
-    const until = new Date("2016-07-01");
+
+  it('computes the balance until a certain month', (done) => {
+    const until = new Date('2016-07-01');
     group.getBalance(until).then(balance => {
       let sum = 0;
       transactions.map(t => {
@@ -95,7 +95,7 @@ describe("Group model", () => {
     })
   });
 
-  it("computes the number of backers", (done) => {
+  it('computes the number of backers', (done) => {
     group.getBackersCount()
       .then(count => {
         expect(count).to.equal(users.length);
@@ -103,8 +103,8 @@ describe("Group model", () => {
       })
   });
 
-  it("computes the number of backers until a certain month", (done) => {
-    const until = new Date("2016-07-01");
+  it('computes the number of backers until a certain month', (done) => {
+    const until = new Date('2016-07-01');
     group.getBackersCount(until).then(count => {
       let backers = {};
       transactions.map(t => {
@@ -116,7 +116,7 @@ describe("Group model", () => {
     })
   });
 
-  it("gets all the expenses", (done) => {
+  it('gets all the expenses', (done) => {
     let totalExpenses = 0;
     transactions.map(t => {
       if(t.netAmountInGroupCurrency < 0)
@@ -128,15 +128,15 @@ describe("Group model", () => {
         expect(expenses.length).to.equal(totalExpenses);
         done();
       })
-      .catch(e => { console.error("error", e, e.stack) });
+      .catch(e => { console.error('error', e, e.stack) });
   });
 
-  it("gets all the expenses in a given month", (done) => {
-    const startDate = new Date("2016-06-01");
-    const endDate = new Date("2016-07-01");
+  it('gets all the expenses in a given month', (done) => {
+    const startDate = new Date('2016-06-01');
+    const endDate = new Date('2016-07-01');
 
     let totalExpenses = 0;
-    
+
     transactions.map(t => {
       if(t.netAmountInGroupCurrency < 0 && t.createdAt > startDate && t.createdAt < endDate)
         totalExpenses++;
@@ -147,6 +147,6 @@ describe("Group model", () => {
         expect(expenses.length).to.equal(totalExpenses);
         done();
       })
-      .catch(e => { console.error("error", e, e.stack) });
+      .catch(e => { console.error('error', e, e.stack) });
   });
 });
